Added servlet context parameter to disable fucntions depending on the server environment. Added the same to the admin panel to hide delete functions

+++ EatSenseRestWS/src/main/java/net/eatsense/restws/AdminResource.java
@@ -1,7 +1,6 @@
 package net.eatsense.restws;
 
 import java.util.ArrayList;
-import java.util.Enumeration;
 import java.util.List;
 
 import javax.servlet.ServletContext;
@@ -12,10 +11,7 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import javax.ws.rs.WebApplicationException;
 
 import net.eatsense.controller.ImportController;
 import net.eatsense.domain.Business;
@@ -29,8 +25,10 @@ import net.eatsense.representation.BusinessImportDTO;
 import net.eatsense.representation.FeedbackFormDTO;
 import net.eatsense.util.DummyDataDumper;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.google.inject.Inject;
-import com.sun.jersey.api.core.ResourceContext;
 
 @Path("admin/services")
 public class AdminResource {
@@ -40,8 +38,8 @@ public class AdminResource {
 	private final ImportController importCtrl;
 
 	private final BusinessRepository businessRepo;
-	private final ServletContext servletContext;
 	protected final Logger logger;
+	private boolean devEnvironment;
 
 	@Inject
 	public AdminResource(ServletContext servletContext, DummyDataDumper ddd,
@@ -49,21 +47,14 @@ public class AdminResource {
 			NicknameAdjectiveRepository adjRepo, NicknameNounRepository nounRepo) {
 		super();
 		this.logger =  LoggerFactory.getLogger(this.getClass());
-		this.servletContext = servletContext;
 		this.ddd = ddd;
 		this.importCtrl = importCtr;
 		this.adjectiveRepo = adjRepo;
 		this.nounRepo = nounRepo;
 		this.businessRepo = businessRepo;
-		String env = servletContext.getInitParameter("karazy.environment");
-		
-		Enumeration params = servletContext.getInitParameterNames();
-		while( params.hasMoreElements() ) {
-			logger.info("initparameter: ", params.nextElement());
-		}
-		
-		logger.info("karazy.enironment: ", env);
-		
+		String environment = servletContext.getInitParameter("net.karazy.environment");
+		// Check for dev environment
+		devEnvironment = environment.equals("dev");
 	}
 	
 	@POST
@@ -130,7 +121,10 @@ public class AdminResource {
 	@DELETE
 	@Path("datastore/all")
 	public void deleteAllData() {
-		importCtrl.deleteAllData();
+		if(devEnvironment)
+			importCtrl.deleteAllData();
+		else
+			throw new WebApplicationException(405);
 	}
 	
 	/**
@@ -139,6 +133,9 @@ public class AdminResource {
 	@DELETE
 	@Path("datastore/live")
 	public void deleteLiveData() {
-		importCtrl.deleteLiveData();
+		if(devEnvironment)
+			importCtrl.deleteLiveData();
+		else
+			throw new WebApplicationException(405);
 	}
 }
